We only need one RowsFieldViewModel typ
