+++ Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/OutputNodeRowViewModel.swift
@@ -14,25 +14,19 @@ final class OutputNodeRowViewModel: NodeRowViewModel {
     static let nodeIO: NodeIO = .output
 
     let id: NodeRowViewModelId
-    
-    
-    // MARK: cached ui-data derived from underlying row observer
-    @MainActor var fieldsUIViewModel: OutputRowFieldsUIViewModel
-    
-//    @MainActor var cachedActiveValue: PortValue
-//    @MainActor var cachedFieldValueGroups = FieldGroupList()
         
-    // MARK: data specific to a draggable port on the canvas; not derived from underlying row observer and not applicable to row view models in the inspector
+    // Cached ui-data derived from underlying row observer
+    @MainActor var fieldsUIViewModel: RowFieldsUIViewModel
+        
+    // Specific to a draggable port on the canvas; not derived from underlying row observer and not applicable to row view models in the inspector
     @MainActor var portUIViewModel: OutputPortUIViewModel
     
-    // MARK: delegates, weak references to parents
-    
+    // Delegates, weak references to parents
     @MainActor weak var nodeDelegate: NodeViewModel?
     @MainActor weak var rowDelegate: OutputNodeRowObserver?
     
     /*
-     // Can an inspector output-row ever have a canvas item delegate ? Or would a "layer output on the graph" be represented as a non-nil canvas item reference on the `OutputLayerNodeRowData` ?
-     // i.e. is this `canvasItemDelegate` only for
+      Can an inspector output-row ever have a canvas item delegate ? Or would a "layer output on the graph" be represented as a non-nil canvas item reference on the `OutputLayerNodeRowData` ?
      */
     @MainActor weak var canvasItemDelegate: CanvasItemViewModel?
