--- Stitch/Graph/Node/Port/ViewModel/NodeRowViewModel/RowFieldsUIViewModel/InputRowFieldsUIViewModel.swift
@@ -1,43 +1,0 @@
-//
-//  InputRowFieldsUIViewModel.swift
-//  Stitch
-//
-//  Created by Christian J Clampitt on 4/22/25.
-//
-
-import Foundation
-
-@Observable
-final class InputRowFieldsUIViewModel: RowFieldsUIViewModel {
-    let id: NodeRowViewModelId
-    
-    @MainActor var cachedActiveValue: PortValue
-    @MainActor var cachedFieldValueGroups: [FieldGroup]
-    
-    @MainActor
-    init(id: NodeRowViewModelId,
-         cachedActiveValue: PortValue,
-         cachedFieldValueGroups: [FieldGroup]) {
-        self.id = id
-        self.cachedActiveValue = cachedActiveValue
-        self.cachedFieldValueGroups = cachedFieldValueGroups
-    }
-}
-
-extension InputNodeRowViewModel {
-    @MainActor var cachedActiveValue: PortValue {
-        get {
-            self.fieldsUIViewModel.cachedActiveValue
-        } set(newValue) {
-            self.fieldsUIViewModel.cachedActiveValue = newValue
-        }
-    }
-    
-    @MainActor var cachedFieldValueGroups: [FieldGroup] {
-        get {
-            self.fieldsUIViewModel.cachedFieldValueGroups
-        } set(newValue) {
-            self.fieldsUIViewModel.cachedFieldValueGroups = newValue
-        }
-    }
-}
